queue: Cleanups, just assume working set fits in RAM for now.
Unfortunately, Save/Load semantics aren't really what we need here,
because it's ok to end up dropping the entire queue on termination.

This still needs to be file backed, for the inevitable time when the
entire mixing queue doesn't fit in RAM, but doing that in a way that's
performant is non-trivial.

Fixes #7.