nonvoting: Add persistence.
Documents and descriptors are now persisted to disk in a boltdb
database.  The load path is untested, but a cursory examination of a
database generated via kimchi makes it seem that things are getting
written correctly at least.

It may be neccecary to GC stale documents and descriptors, but for now
the database is append-only to preserve such things in perpetuity.

Fixes #1.