nonvoting: Clarify the `410` error message.
In *most* cases a `410 Gone` means that the client is requesting a
document that is in the past.  However, the error is actually returned
for all cases when the authority knows that it will NEVER be able to
satisfy the client's request, even if they wait a bit and retry, for
example if insufficient nodes published descriptors for the next epoch.