minclient: Adjust the fetch delay instead of resetting after send.
Fetches should happen independently of sends.  The old code was
resetting the fetch delay back to full, which could result in fetches
never getting executed due to the timer continually getting reset by
consecutive sends.

This adjusts the time till the next fetch after a send to resume the
delay with the correct (reduced) interval.

Fixes #6.