nonvoting: Use the wire protocol to talk to the authority.
This uses EdDSA -> X25519 key conversion which isn't great.  A real
authority probably should bite the bullet and either sign link keys
somehow, or distribute a link key along with the signing key.

That said, it's a lot better than HTTP, and I'm fairly sure doing so is
safe in this context.

WARNING: This is a backwards incompatible change, because mixes,
providers, the authority and clients all need to use the same
protocol to communicate.

Fixes #12.