provider: Optionally allow multiple provider workers.
The current user message spool implementation renders this largely
pointless as the bulk of the provider worker's processing time is going
to be occupied with database transactions.  At some point in the future
a concurrent user message spool implementation will need to be written
and this lays dow the groundwork for properly utilizing such.

The associated tunable is `Debug.NumProviderWorkers`, which will default
to 1 for the time being.

Part of #4.