provider: Support a RDBMS backend.
This adds support for interfacing with an RDBMS, where "RDBMS" is
defined as Postgresql.  The Spool `Vaccum()` method is not supported for
cases where the RDBMS is only used as the spool (and not the UserDB).
For now it is assumed that those who wish to use such a configuration
are capable of manually cleaning up the various tables.

Fixes #4.