Make it work for NATed environments
Basically this makes it possible to advertise an alternate address only
such that it can use an internal transport such as TCP4 or TCP6 but
it does NOT advertise the address it is actually listening on. This makes
it possible for the mix server to operate behind a NAT device where it's
internal IP address differs from it's external or public facing IP address.