return to stateBootstrap and retry consensus fetch
If the authority waits until the next descriptor submission deadline,
it may miss retrieving the consensus made in the previous epoch and will
not have a copy of the previous consensus from which to inherit the
correct topology. This commit fixes this issue by returning to
stateBootstrap and retrying a consensus fetch for the current epoch.