Adds TimerQ type and unit tests
The TimerQ is a container that holds messages until an expiry time, upon
which it will pass the message via an interface that has one method,
Push.

Each message is placed into a priority queue ordered by earliest
message expiry. A single time.Timer is used to wake the worker thread
for the message expiry events. The worker thread is also woken upon
Enqueue and Remove method calls in order to update the timer.

TODO: the message expiry function should be corrected/improved. Possibly
it should be passed as a constructor argument.