Fix TimerQ, use priority properly
Here we make sure to not complect and add
dependecies on Message fields. That is, for the
entry priority we need not lookat Message fields
because the Priority Queue's Entry type has
the Priority field which should be used.

Also since this will be used for many different purposes it is
not logical to assume the priority will always be set by Message's
ReplyETA field.