do not register and then not use a linkkey
catchat registers a new linkkey, but does not actually use it.
the desired behavior is actually to rotate linkkeys when changing
network connectivity, OR, on every startup. however, this behavior
prevents offline usage of the client - which requires a lot of
refactoring to support anyway, so this is marked as a TODO.