round timestamp to fix suspend bug
Prior to this commit, 10h old messages would claim to be 2h old after my
computer was asleep for 8h. It turns out this is a feature of golang:

"Rather than split the API, in this package the Time returned by time.Now
contains both a wall clock reading and a monotonic clock reading; later
time-telling operations use the wall clock reading, but later time-measuring
operations, specifically comparisons and subtractions, use the monotonic clock
reading."

"On some systems the monotonic clock will stop if the computer goes to sleep.
On such a system, t.Sub(u) may not accurately reflect the actual time that
passed between t and u."

"The canonical way to strip a monotonic clock reading is to use t = t.Round(0)."

https://pkg.go.dev/time
https://go.googlesource.com/proposal/+/master/design/12914-monotonic.md

We should probably think about all of our other uses of golang's time objects
and which clock we actually want to be using; there may be other
suspend-related bugs.