fix display glitch by serializing microseconds
Prior to this commit, if one sent and received messages during the same second,
those two messages would lose their canonical sort order after the client was
relaunched, which in turn would cause the messages to frequently swap their
order on the screen during redraws.

This is because the default Time encoding option in the cbor library we're
using is cbor.TimeUnix which stores time as an integer number of seconds:

https://pkg.go.dev/github.com/fxamacker/cbor/v2#readme-encoding-options-1

Conveniently, there is a forward-compatible Time encoding option called
cbor.TimeUnixDynamic which will accept the old low-precision timestamps while
storing new ones with microsecond precision. So, this commit does that, which
fixes the display glitch problem for new messages.

Note that due to the way the cbor library works this change to the state file
is actually both forward *and* backwards compatible. However, going back to an
older version will cause all stored stampstamps to lose their microsecond
precision.