Add work-in-progress "Control Port" server
Here's my "Katzenpost client control port".
The server (in this case the katzenpost client is the control port server) streams
events to the client. While the client streams commands to the server. However
this is not a strict call and response protocol. It's simultaneous bidirectional communication.
Uses length prefixed CBOR.