Adds AuthenticationType to MixDescriptor
Adds a field called AuthenticationType to the MixDescriptor. This can be
used to indicate what type of authentication is required to use a
provider - if a provider has EnableEphemeralClients and TrustOnFirstUse
enabled, then the AuthenticationType string is set to "Anonymous". A
helper method NewEphemeralClient is added to client in order to select a
compatible provider.