docker/nonvoting_mixnet: use testnet parameters
this patch was copied from 8d27565363b9b which did this for the voting mixnet
on the main branch, which is now diverged from here. I'm adding this here in
preparation for using this branch to test changes which have already landed on
main but are not yet deployed to the testnet.