ping: send all pings in parallel
note: it would be better if this was optional behavior, and if it handled a
full queue. currently this works up to n=40 but above that it gets "queue is
full" errors. it would also nice to print packets-per-second stats which ignore
the time spent connecting to the network and fetching the consensus.