eddsa/blinded25519: the Tor V3 onion key blinding scheme for ed25519
This is a port of the scheme used by Tor. It will be useful as a building block
for features that are supposed to be unlinkable-to-some-but-not-everybody, where
we can use it to derive spool IDs and other identifiers asynchronously.

The PR also touches on eddsa, removing the dependency on extra25519 which is now only used in the `core/crypto/reunion/elligator.go` module. Once gone from there, we can remove the old vendored `edwards25519` and `extra25519` modules.