docker: workaround port mapping
this is a workaround that enables nested containerization. katzenpost
needs an ip address specified in its configuration file, and podman does
not portmap services listening on loopback in containers. podman-compose
does not seem to support creating the routing rules to map services to
the ip_address specified in the configuration yaml, so binding to all
interfaces, but announcing the loopback address, and using host
networking mode means that all the mixes / authorities can communicate
and ports mapped outside the container listen on localhost so the PKI
documents are also correct (entries contain localhost and not an
unroutable internal network...)

I think a better solution would be to figure out how to enter the
network namespace of the container so that an application like katzen
can be used with the containerized mixnet.

or, to figure out how to use an external network directive that is
routed between the host and container

but this workaround does work for now.