fix for half of #133
this hopefully fixes the first of the two tracebacks in issue #133:

    panic: interface conversion: interface {} is *decoy.surbCtx, not []*decoy.surbCtx)

apparently the avl library returns a slice of pointers except for sometimes
when it just returns a single item.

it remains a mystery to me why we just started seeing this panic recently, and
(afaik) on only one mixnode.

(this commit written while talking to @david415)