Stream: Add Dial() and Listen() methods
NewStream now generates keymaterial as a listener by default.
Peers can use the LocalAddr() to Dial() and establish keymaterial as dialer.

TODO: keymaterial should be the output of a DH handshake between peers
such that the message storage IDs are blinding factors of a the public
keys and reader/writer capabilities can be implemented