Stream: fix more off-by-one errors calculating Acknowledgements
s.ReadIdx points at the peer's next frame, not the last successfully fetched frame.
So s.ReadIdx > s.WindowSize is correct and not >=, and also at Teardown the Ack should
correspond to the last successfully read frame (not ReadIdx but ReadIdx - 1).

There is a special case of a sender having read 0 frames, so frame.Ack == 0 is a special case
and should probably be handled in a better way.