Improved Sphinx API design
1. The Sphinx Geometry type functions as a factory for creating
our Sphinx type which knows how to create and unwrap Sphinx packets.
No other information other than the Geometry type will be needed.

2. The Sphinx Geometry fully serializes easily... as it's composed
of ints and strings, actually two strings, the name of the KEM and
the name of the NIKE where one of them must be an empty string.