server: Remove AltAddresses in favor of URL Address lines
This gets rid of the AltAddresses config option in favor of adding a
protocol scheme to each Address entry (as a URL).

In order to Advertise an Address that the server is not bound to (for
example, due to NAT and port forwarding) the configuration will support
an option OnlyAdvertiseAddresses which will control which listeners are
published in the PKI. This isn't implemented yet, nor do the Authorities
validate that the Addresses are actually valid.

Note - WIP: this SHOULD work but for some reason the listener isn't receiving
messages and times out.