map/client: Replace PayloadSize with PayloadSize()
As the sphinx Geometry is defined at runtime, we replace the package variable with a method that uses the Geometry from the Session.
Care must be used to ensure that Payload size changes do not break any protocols, i.e. Payload size changes at runtime must be detected and handled.