Sockatz: start extending client listener to support UDP Associate
The SOCKS5 protocol described in rfc1928 (https://datatracker.ietf.org/doc/html/rfc1928)
specifies that a client can send a UDP Associate command that the SOCKS5 server can reply to with the address:port of a listener that UDP datagrams should be sent to.

This extends Sockatz to support reading the UDP Associate commands and open a listening UDP Connection.

To be implemented:
 * Handlers that process incoming UDP packets on each listening UDPConn, verifying that only UDP packets for the requested associated hosts are proxied to the Sockatz Kaetzchen service
 * Parser for the UDP packet header format (defined in rfc1928)
 * Mapping from incoming TCP connection to Associated UDPConns so that upon TCPConn close the UDP listeners are terminated.
 * Sockatz Kaetzchen service UDP connection worker