Fix transient errors in TestBadCertificate: ```  --- PASS: TestWrongCertificate (0.92s) === NAME  TestBadCertificate     cert_test.go:87:          Error Trace: /home/runner/work/katzenpost/katzenpost/core/crypto/cert/cert_test.go:87          Error:       An error is expected but got nil.          Test:        TestBadCertificate --- FAIL: TestBadCertificate (0.61s) ```
This happens when `certificate[1000] == 235` in which case no modification is made,
and we unexpectedly get an OK from `Verify()`. This used to happen in 1/256 test executions.

Also bumps the expiry by one so we're not impacted by epoch rollover between `Sign()` and `Verify()`,
which we haven't run into, but could happen one day.