sockatz: use nonblocking SendMessage API
This adds an additional worker routine to process incoming messages
asynchronously from the send worker.

SendUnreliableMessage returns an error if the egress queue is full,
so we need to backoff retrying until the queue has emptied.

Also increase the size of the client_test transferred payload, because
this is a lot faster now.