fix testnet shell
i broke this 13 months ago... the shell target should be able to get online. it
was not because of a stray -e argument (to set environment) which consumed the
--network=host argument.

this also removes the --network=host (and stray -e) from the ping build command,
since it doesn't need network at all.