testnet: cache build artifacts
With this commit, the /go and /root/.cache directories are mounted from outside
the containers (in ./docker/cache) to make building *much* faster.

This is especially necessary since highctidh was making each of our many
binaries that link against it take minutes to build. With the build cache, all
after the first one take mere seconds again.

The pipe syntax in the Makefile designates "order-only-prerequisites", so that
targets which depend on $(cache_dir) won't be re-run just because the mtime of
the cache dir is newer than they are. See
https://www.gnu.org/software/make/manual/html_node/Prerequisite-Types.html for
details.

This commit also removes the katzencat make target (as katzencat hasn't been
merged yet).