revert "catshadow: add ratchet recv key assertion for send"
this reverts this, plus my patches to backport it to v30:
    commit 02962240d48ff146a67a6f342829c3682c4a01c7
    Author: David Stainton <dstainton415@gmail.com>
    Date:   Sat Jun 15 21:25:46 2024 -0700

        catshadow: add ratchet recv key assertion for send

fascinatingly having this assertion makes the UpgradeResume test fail in v30,
despite that the test passes without this Assertion (more messages are possible
after reloading the saved state files that are generated here on the v30 branch
and saved in git). I am reverting this to confirm that it is still passing
without this change.