fix the doubleratchet upgrade bug 🥳
the bug was that the nike doubleratchet swapped the position of the nonce and
the public keys.

this fixes that so that (hopefully) the nike doubleratchet using
NOBSCSIDH512x25519 is now compatible with the old hybrid ratchet in katzenpost
v0.0.30 - even if there are messages in flight during the upgrade, or if the
two parties upgrade at different times.