catshadow: use TimeUnixMicro for better than 1s resolution
catshadow sorts messages by Timestamp. This workaround tells the cbor library to use the option TimeMicroUnix instead of TimeUnix.
Because messages are encrypted with the senders timestamp before queuing for transmission, it's easily possible for a user to send more than 1 message per second, which is the default timestamp resolution, and results in unstable sort order of messages.