core/wire: increase max msg len to 40MB
MaxMsgLen does NOT dictate the message size, instead... our padding scheme upon
initialization will use the largest commands to set the size of the padding.

Get rid of MaxMsgLen?
No way. it's useful to have it as a sanity/validity check against the untrusted data we receive
on the network.

It turns out that the right solution here is to set MaxMsgLen to a sufficiently high value so that
we can select any KEM for Sphinx.... this in turn means a multiplier and huge growth in size for the
dirauth voting command which includes all the descriptors and their mix keys. If the mix keys go from
32 bytes to 270 bytes... it's effects a huge difference in sizes for the voting and the final PKI documents.
These huge binary blobs must fit into a single Noise message.