replica/courier: fix dependency hierarchy, fixes CI tests
when the replica is sharing some of it's code with the courier, we don't
want the courier to be required to have the rocksdb dependency. Therefore I moved
the ReplicaMap type (and it's methods) over to the replica/common/ subdir so that when
it's imported from the courier, it works even if the courier is running on a machine
that does not have rocksdb-dev installed.