fixup ci workflow
um temporarily disable race detection on the courier unit tests...
later when we fix the tests so that it's not sleeping and the waking up to check state,
and instead communicating over channels to block and wake up threads etc. then at that time,
we can re-enable race detction on these tests. but the tests work and yes
they have races that's kind of the point of the sleeps, causing our tester thread to lose said races ;-)