Ensure kpclientd leaves local socket open
When a thinclient disconnects from kpclientd, we want to keep the
listener socket around for future thinclient usage. This change
suppresses an errno CBOR serialization issue by returning nil rather
than the real error status. This change strategy prevents needing to
update the thinclient for now and allows the python thinclient to reuse
the socket without issues. It additionally resolves a goroutine leak.