Parallel Construction (of Descriptor Uploads)
An operator noticed that their mix server descriptor was being posted in
sequence with absurdly long time outs. This caused a pathological
failure mode.

In authority/voting/client.go we refactored to run in parallel rather
than sequentially and removed a sleep to allow allow canceling. This
maybe fixes a connection leak by adding a defer. Tests added to cover
parallism.

In core/retry.go we set some halfbaked defaults that need to be tuned:
- DefaultMaxAttempts: 10 -> 20
- DefaultBaseDelay: 500ms -> 10ms
- DefaultMaxDelay: 10s -> 3s